backend/pyproject.toml
[tool.poetry]
name = "portfolio"
version = "0.1.0"
description = ""
authors = ["hojiakbar2806 <hojiakbarnasriddinov2006@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
asyncpg = "^0.30.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.37"}
alembic = "^1.14.1"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
